requirements.txt
beautifulsoup4
certifi
dnspython
Pillow
pytesseract
requests
tldextract
whois
